refactor(editorial): extract editorial cells, pack, and pipeline into extensions/editorial (#2217)
