+++ extensions/editorial/pack/agents/editorial-draft.json
