+++ extensions/editorial/pack/agents/editorial-draft.md
