+++ extensions/editorial/pack/agents/editorial-research.json
