+++ extensions/editorial/pack/agents/editorial-research.md
