+++ extensions/editorial/pack/pack.json
@@ -1,0 +1,5 @@
+{
+  "name": "editorial",
+  "version": "1.0.0",
+  "namespace": "editorial"
+}
