+++ extensions/editorial/pack/schemas/editorial-draft.input.json
