+++ extensions/editorial/pack/schemas/editorial-research.input.json
